u64 shuffle

+++ inc/rand.h
@@ -9,5 +9,5 @@ void reset_seed(void);
 void rand_seed(uint64_t seed);
 uint64_t randn(void);
 uint64_t randrange(uint64_t low, uint64_t hi);
-void shuffle(void* arr, int szof, int n);
+void shuffle(void* arr, uint64_t szof, uint64_t n);
 #endif /* RAND_H */
